pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "javs"
version = "1.0.0"
description = "A powerful CLI tool to scrape, organize, and manage JAV media libraries"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    "typer[all]>=0.9",
    "pydantic>=2.0",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "tenacity>=8.0",
    "structlog>=23.0",
    "aiofiles>=23.0",
    "cloudscraper>=1.2",
]

[project.optional-dependencies]
translate = ["googletrans==4.0.0-rc.1", "deepl>=1.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "aioresponses>=0.7",
]

[project.scripts]
javs = "javs.cli:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
